Refactor schedule related files to be in separate package
