+++ app/src/main/java/com/github/sdpcoachme/DashboardActivity.kt
@@ -41,6 +41,7 @@ import com.github.sdpcoachme.DashboardActivity.TestTags.Companion.MENU_LIST
 import com.github.sdpcoachme.data.MapState
 import com.github.sdpcoachme.errorhandling.ErrorHandlerLauncher
 import com.github.sdpcoachme.map.MapViewModel
+import com.github.sdpcoachme.schedule.ScheduleActivity
 import com.github.sdpcoachme.ui.MapView
 import com.github.sdpcoachme.ui.theme.CoachMeTheme
 import com.google.android.gms.location.FusedLocationProviderClient
